+++ packages/api/src/controllers/tokenController.ts
@@ -1,9 +1,10 @@
 import { getRepository } from 'typeorm';
-import { Controller, Get, Req, HttpError } from 'routing-controllers';
+import { Controller, Get, Req, UnauthorizedError } from 'routing-controllers';
 import { authService } from 'src/services/authService';
 import { User } from 'src/entities/user';
 import { verify } from 'jsonwebtoken';
 import { Request } from 'express';
+import { ERROR_MESSAGES } from 'src/utils/constants';
 
 @Controller('/token')
 export class tokenController {
@@ -13,16 +14,14 @@ export class tokenController {
   async sendAccessToken(@Req() request: Request) {
     const refreshToken = request.headers['refresh-token'] as string;
     if (typeof refreshToken !== 'string')
-      throw new HttpError(401, 'refresh token not found !');
+      throw new UnauthorizedError(ERROR_MESSAGES.refresh_token_not_found);
     const token = refreshToken.split('Bearer ')[1];
-    if (!token) throw new HttpError(401, 'refresh token not found !');
+    if (!token) throw new UnauthorizedError(ERROR_MESSAGES.refresh_token_not_found);
     let data;
     try {
-      data = <{ userId: string }>(
-        verify(token, process.env.REFRESH_TOKEN_SECRET as string)
-      );
+      data = <{ userId: string }>verify(token, process.env.REFRESH_TOKEN_SECRET as string);
     } catch {
-      throw new HttpError(401, 'Token is expired or invalid !');
+      throw new UnauthorizedError(ERROR_MESSAGES.acccess_token_expired);
     }
     const user = await this.userRepo.findOne({ id: data.userId });
     return this.authService.createTokens(user as User);