+++ packages/api/src/controllers/userController.ts
@@ -1,46 +1,60 @@
-import { Request, Response } from 'express';
-import { verify } from 'jsonwebtoken';
-import { Controller, Get, HttpError, Req, Res } from 'routing-controllers';
+import { Response } from 'express';
+import {
+  BadRequestError,
+  Body,
+  Controller,
+  Get,
+  InternalServerError,
+  Patch,
+  Req,
+  Res,
+  UseBefore,
+} from 'routing-controllers';
 import { User } from 'src/entities/user';
+import { authMiddleware, RequestWithUser } from 'src/middlewares/auth.middleware';
 import { authService } from 'src/services/authService';
+import { ERROR_MESSAGES } from 'src/utils/constants';
 import { getRepository } from 'typeorm';
 
+@UseBefore(authMiddleware)
 @Controller('/user')
 export class userController {
   private readonly userRepo = getRepository(User);
   private readonly authService = new authService();
+
   @Get('/')
-  async getUserFromToken(@Req() request: Request, @Res() response: Response) {
-    const accessToken = request.headers['access-token'] as string;
-    if (typeof accessToken !== 'string')
-      return response
-        .status(401)
-        .send(new HttpError(401, 'access token not found !'));
-    const token = accessToken.split('Bearer ')[1];
-    if (!token)
-      return response
-        .status(401)
-        .send(new HttpError(401, 'access token not found !'));
-    let data: any;
-    try {
-      data = <{ userId: string }>(
-        verify(token, process.env.ACCESS_TOKEN_SECRET as string)
-      );
-    } catch {
-      return response
-        .status(401)
-        .send(new HttpError(401, 'Token is expired or invalid !'));
-    }
-    const user = await this.userRepo.findOne({ id: data.userId });
-    if (!user)
-      return response
-        .status(400)
-        .send(new HttpError(400, "user doesn't exist !"));
+  async getUserFromToken(@Req() { userId }: RequestWithUser, @Res() response: Response) {
+    const user = await this.userRepo.findOne({ id: userId });
+    if (!user) return response.status(400).send(ERROR_MESSAGES.user_not_found);
 
     (user.password as any) = undefined;
+    (user.api_key as any) = undefined;
     return {
       ...user,
       ...this.authService.createTokens(user),
     };
+  }
+
+  @Patch('/')
+  async updateUser(@Req() { userId }: RequestWithUser, @Body() user: Partial<User>) {
+    try {
+      const userFromDb = await this.userRepo.findOne({ where: { id: userId } });
+      if (!userFromDb) throw new BadRequestError(ERROR_MESSAGES.user_not_found);
+      return await this.userRepo.update({ id: userId }, { ...user });
+    } catch (error) {
+      throw new BadRequestError(ERROR_MESSAGES.user_exists_with_credentials);
+    }
+  }
+  @Get('/key')
+  async getApiKey(@Req() { userId }: RequestWithUser) {
+    try {
+      const userFromDb = await this.userRepo.findOne({ where: { id: userId } });
+      if (!userFromDb) throw new BadRequestError(ERROR_MESSAGES.user_not_found);
+      const { api_key } = userFromDb;
+      if (api_key) return api_key;
+      return await this.authService.generateApiKey(userId);
+    } catch (error) {
+      throw new InternalServerError(ERROR_MESSAGES.iss);
+    }
   }
 }