+++ packages/app/src/components/AppModal.tsx
@@ -4,20 +4,16 @@ interface Props {
   isModal: boolean;
   closeModal: () => void;
   heading: string;
+  subheading?: string;
 }
 
-const AppModal: React.FC<Props> = ({
-  isModal,
-  closeModal,
-  heading,
-  children,
-}) => {
+const AppModal: React.FC<Props> = ({ isModal, closeModal, heading, children, subheading }) => {
   return (
     <div>
       <div
         className={`modal ${
           !isModal ? 'opacity-0 pointer-events-none' : ''
-        }  fixed w-full h-full top-0 left-0 flex items-center justify-center`}
+        }  fixed w-full h-full top-0 left-0 flex items-center justify-center z-40`}
       >
         <div
           className="modal-overlay absolute w-full h-full bg-gray-900 opacity-50"
@@ -27,11 +23,13 @@ const AppModal: React.FC<Props> = ({
         <div className="modal-container bg-white w-11/12 md:max-w-2xl mx-auto rounded-md shadow-lg z-50 overflow-y-auto">
           <div className="modal-content py-4 text-left px-6">
             <div className="flex justify-between items-center pb-3 ">
-              <p className="text-2xl font-bold">{heading}</p>
-              <div
-                className="modal-close cursor-pointer z-50"
-                onClick={() => closeModal()}
-              >
+              <div>
+                {' '}
+                <p className="text-2xl font-bold">{heading}</p>
+                <p className="text-sm text-gray-500 "> {subheading} </p>
+              </div>
+
+              <div className="modal-close cursor-pointer z-50" onClick={() => closeModal()}>
                 <svg
                   className="fill-current text-black"
                   xmlns="http://www.w3.org/2000/svg"