+++ packages/app/src/store/userContext.ts
@@ -1,3 +1,5 @@
+import { AxiosResponse } from 'axios';
+import { DonkeyApi } from 'src/utils/helpers';
 import create, { State } from 'zustand';
 
 export interface User {
@@ -15,6 +17,7 @@ export interface UserState extends State {
   user: User;
   setLogin: (load: boolean) => void;
   setUser: (load: User) => void;
+  fetchUser: () => Promise<any>;
 }
 
 export const useUser = create<UserState>((set, get) => ({
@@ -22,4 +25,9 @@ export const useUser = create<UserState>((set, get) => ({
   user: {},
   setLogin: (load) => set(() => ({ isLoggedIn: load })),
   setUser: (load) => set(() => ({ user: { ...load } })),
+  fetchUser: async () => {
+    await DonkeyApi.get('/user').then((res: AxiosResponse<User>) =>
+      set(() => ({ user: { ...res.data } })),
+    );
+  },
 }));